reorganize methods to locate in suitable files
