+++ app/src/main/java/com/example/categorynoteapp/FragmentUtils.kt
@@ -1,7 +1,9 @@
 package com.example.categorynoteapp
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.commit
 import com.example.categorynoteapp.ui.category.CategoryFragment
 
 fun setupFirstFragmentIfNotSetup(savedInstanceState: Bundle?, supportFragmentManager: FragmentManager) {
@@ -12,4 +14,13 @@ fun setupFirstFragmentIfNotSetup(savedInstanceState: Bundle?, supportFragmentMan
 
 fun setupFirstFragment(supportFragmentManager: FragmentManager) {
     supportFragmentManager.beginTransaction().add(R.id.container, CategoryFragment()).commit()
+}
+
+fun Fragment.changeFragment(args: Bundle, fragmentManager: FragmentManager) {
+    this.arguments = args
+
+    fragmentManager.commit {
+        addToBackStack(null)
+        replace(R.id.container, this@changeFragment)
+    }
 }
